fix bug happening when creating a global client
